Started refactoring Invoice views

+++ WMgr/Models/InvoiceDetail.cs
@@ -11,7 +11,7 @@ namespace WMgr.Models
         public int InvoiceDetailId { get; set; }
         public virtual int InvoiceId { get; set; } //link to invoice header
 
-        public float Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public virtual Product Product { get; set; }
         public decimal SellPrice { get; set; }
     }
